requirements.txt
kaleido==0.2.1
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
plotly==5.13.1
pytorch-lightning==1.9.4
SciencePlots==2.0.1
scikit-learn==1.2.1
scipy==1.10.1
statsmodels==0.14.0
torch==1.13.1
tqdm==4.65.0
wandb==0.13.10
PyYAML==6.0.1